+++ backend/history/activity_scheduled.go
@@ -2,7 +2,7 @@ package history
 
 import (
 	"github.com/cschleiden/go-workflows/backend/metadata"
-	"github.com/cschleiden/go-workflows/internal/payload"
+	"github.com/cschleiden/go-workflows/backend/payload"
 )
 
 type ActivityScheduledAttributes struct {
