+++ backend/history/side_effect_result.go
@@ -1,6 +1,6 @@
 package history
 
-import "github.com/cschleiden/go-workflows/internal/payload"
+import "github.com/cschleiden/go-workflows/backend/payload"
 
 type SideEffectResultAttributes struct {
 	Result payload.Payload `json:"result,omitempty"`
