+++ backend/history/subworkflow_completed.go
@@ -1,6 +1,6 @@
 package history
 
-import "github.com/cschleiden/go-workflows/internal/payload"
+import "github.com/cschleiden/go-workflows/backend/payload"
 
 type SubWorkflowCompletedAttributes struct {
 	Result payload.Payload `json:"result,omitempty"`
