+++ backend/history/workflow_completed.go
@@ -1,7 +1,7 @@
 package history
 
 import (
-	"github.com/cschleiden/go-workflows/internal/payload"
+	"github.com/cschleiden/go-workflows/backend/payload"
 	"github.com/cschleiden/go-workflows/internal/workflowerrors"
 )
