+++ backend/history/workflow_continued_as_new.go
@@ -1,6 +1,6 @@
 package history
 
-import "github.com/cschleiden/go-workflows/internal/payload"
+import "github.com/cschleiden/go-workflows/backend/payload"
 
 type ExecutionContinuedAsNewAttributes struct {
 	Result payload.Payload `json:"result,omitempty"`
