+++ internal/args/args_test.go
@@ -5,8 +5,8 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/cschleiden/go-workflows/backend/payload"
 	"github.com/cschleiden/go-workflows/converter"
-	"github.com/cschleiden/go-workflows/internal/payload"
 	"github.com/stretchr/testify/require"
 )
