+++ internal/command/complete_workflow_test.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/benbjohnson/clock"
 	"github.com/cschleiden/go-workflows/backend/history"
+	"github.com/cschleiden/go-workflows/backend/payload"
 	"github.com/cschleiden/go-workflows/core"
-	"github.com/cschleiden/go-workflows/internal/payload"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 )
