+++ internal/command/sideeffect.go
@@ -3,7 +3,7 @@ package command
 import (
 	"github.com/benbjohnson/clock"
 	"github.com/cschleiden/go-workflows/backend/history"
-	"github.com/cschleiden/go-workflows/internal/payload"
+	"github.com/cschleiden/go-workflows/backend/payload"
 )
 
 type SideEffectCommand struct {
