+++ internal/workflowstate/signalchannels.go
@@ -1,8 +1,8 @@
 package workflowstate
 
 import (
+	"github.com/cschleiden/go-workflows/backend/payload"
 	"github.com/cschleiden/go-workflows/internal/converter"
-	"github.com/cschleiden/go-workflows/internal/payload"
 	"github.com/cschleiden/go-workflows/internal/sync"
 )
