+++ internal/workflowstate/workflowstate_test.go
@@ -5,9 +5,9 @@ import (
 	"testing"
 
 	"github.com/benbjohnson/clock"
+	"github.com/cschleiden/go-workflows/backend/payload"
 	"github.com/cschleiden/go-workflows/converter"
 	"github.com/cschleiden/go-workflows/core"
-	"github.com/cschleiden/go-workflows/internal/payload"
 	"github.com/cschleiden/go-workflows/internal/sync"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
